requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
base58==2.1.1
beautifulsoup4==4.12.3
bitarray==3.0.0
blinker==1.9.0
cachetools==5.5.1
certifi==2024.12.14
charset-normalizer==3.4.1
ckzg==2.0.1
click==8.1.8
construct==2.10.68
construct-typing==0.5.6
cytoolz==1.0.1
dataclasses-json==0.6.7
distro==1.9.0
eth-account==0.13.4
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.6.1
eth-rlp==2.1.0
eth-typing==5.1.0
eth-utils==5.2.0
eth_abi==5.2.0
frozenlist==1.5.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.14.0
hexbytes==1.3.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
Jinja2==3.1.5
jiter==0.8.2
jsonalias==0.1.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain==0.3.15
langchain-community==0.3.15
langchain-core==0.3.31
langchain-ollama==0.2.2
langchain-openai==0.3.2
langchain-text-splitters==0.3.5
langserve==0.3.1
langsmith==0.3.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.0
mdurl==0.1.2
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.23.0
numpy==2.2.2
ollama==0.4.7
openai==1.60.1
orjson==3.10.15
packaging==24.2
pandas==2.2.3
parsimonious==0.10.0
pillow==11.1.0
propcache==0.2.1
protobuf==5.29.3
pyarrow==19.0.0
pycryptodome==3.21.0
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pyunormalize==16.0.0
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rlp==4.0.1
rpds-py==0.22.3
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
solana==0.36.3
solders==0.23.0
soupsieve==2.6
SQLAlchemy==2.0.37
streamlit==1.41.1
tenacity==9.0.0
tiktoken==0.8.0
toml==0.10.2
toolz==1.0.0
tornado==6.4.2
tqdm==4.67.1
types-requests==2.32.0.20241016
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
web3==7.7.0
websockets==13.1
yarl==1.18.3
zstandard==0.23.0
